% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_is.R
\name{as_is}
\alias{as_is}
\title{as_is}
\usage{
as_is(composition, dm, digits = 2)
}
\arguments{
\item{composition}{feed chemical composition (dry-matter basis).}

\item{dm}{percentage of dry matter contents (without moisture) of feed.}

\item{digits}{integer indicating the number of decimal places}
}
\description{
This function convert dry-matter basis into as-is basis.
}
\examples{
df <- data.frame(nutrient=c("CP","NDF","EE","Ash"),composition=c(10,30,5,2))
as_is(df$composition,85)
}
\keyword{feed}
\keyword{formulation}
