% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_m.R
\name{d_m}
\alias{d_m}
\title{d_m}
\usage{
d_m(composition, dm, digits = 2)
}
\arguments{
\item{composition}{feed chemical composition (as-is basis).}

\item{dm}{percentage of dry matter contents (without moisture) of feed.}

\item{digits}{integer indicating the number of decimal places}
}
\description{
This function convert as-is basis into dry-matter basis.
}
\examples{
df <- data.frame(nutrient=c("CP","NDF","EE","Ash"),composition=c(10,30,5,2))
d_m(df$composition,85)
}
\keyword{feed}
\keyword{formulation}
