% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.extract.R
\name{feed.extract}
\alias{feed.extract}
\title{feed.extract}
\usage{
feed.extract(url, n)
}
\arguments{
\item{url}{A URL that you want to scraping.}

\item{n}{A number of list from feed.list().}
}
\description{
This function extract and re-combine the list from feed.list().
}
\examples{
feed.extract("https://www.feedipedia.org/node/556",2)
}
\keyword{feed,}
\keyword{feedipedia}
