% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.info.R
\name{feed.info}
\alias{feed.info}
\title{feed.info}
\usage{
feed.info(url)
}
\arguments{
\item{url}{A URL that you want to scraping.}
}
\description{
This function show the result of tables of URL.
}
\examples{
feed.info("https://www.feedipedia.org/node/556")
}
\keyword{feed,}
\keyword{feedipedia}
