% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.list.R
\name{feed.list}
\alias{feed.list}
\title{feed.list}
\usage{
feed.list(category = "https://www.feedipedia.org/content/feeds?category=All")
}
\arguments{
\item{category}{A category that you want to find(all; forage; plantProduct; animalOrigin; other)}
}
\description{
This function showed a feed list at feedipedia.org.
}
\examples{
feed.list(all)
feed.list(forage)
feed.list(plantProduct)
feed.list(animalOrigin)
feed.list(other)
}
\keyword{feed,}
\keyword{feedipedia}
