% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.nrc.beef.R
\docType{data}
\name{feed.nrc.beef}
\alias{feed.nrc.beef}
\title{Feed chemical composition from NRC beef 2016}
\format{A data frame with 219 rows and 57 variables:}
\usage{
feed.nrc.beef
}
\description{
Feed chemical composition from NRC beef 2016
}
\keyword{datasets}
