% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.search.R
\name{feed.search}
\alias{feed.search}
\title{feed.search}
\usage{
feed.search(keyword)
}
\arguments{
\item{keyword}{A keyword that you want to find.}
}
\description{
This function search a keyword at feedipedia.org.
}
\examples{
feed.search("alfalfa")
}
\keyword{feed,}
\keyword{feedipedia}
