% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.compare.R
\name{feed.compare}
\alias{feed.compare}
\title{feed.compare}
\usage{
feed.compare(feed1, feed2, name1 = "first", name2 = "second",
  radar = TRUE)
}
\arguments{
\item{feed1}{1st feed chemical composition that you want to compare.}

\item{feed2}{2nd feed chemical composition that you want to compare.}

\item{name1}{1st feed name that you want to compare.}

\item{radar}{TRUE returns radar chart.}

\item{name1}{2nd feed name that you want to compare.}
}
\description{
A function to compare each result of feed composition.
}
\examples{
feed1 <- feed::feed.info("https://www.feedipedia.org/node/556",2)
feed2 <- feed::feed.info("https://www.feedipedia.org/node/674",3)
feed1 <- feed1[[1]]
feed2 <- feed2[[1]]
feed.compare(feed1,feed2,name1="corn",name2="sbm",radar=TRUE)
}
\keyword{feed}
\keyword{formulation}
