% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.list.R
\name{feed.list}
\alias{feed.list}
\title{feed.search}
\usage{
feed.list(url)
}
\arguments{
\item{url}{A URL that you want to scraping.}
}
\description{
This function show the result of tables of URL.
}
\examples{
feed.list("https://www.feedipedia.org/node/556")
}
\keyword{feed,}
\keyword{feedipedia}
